Build: Moving JIT and test cleaning utilities to loki.build.jit
