+++ loki/tests/test_expression.py
@@ -14,9 +14,8 @@ import numpy as np
 
 import pymbolic.primitives as pmbl
 
-from conftest import (
-    jit_compile, clean_test, stdchannel_redirected, stdchannel_is_captured
-)
+from conftest import stdchannel_redirected, stdchannel_is_captured
+
 from loki import (
     Sourcefile, cgen, fgen, Cast, RangeIndex, Assignment, Intrinsic, Variable,
     Nullify, IntLiteral, FloatLiteral, IntrinsicLiteral, InlineCall, Subroutine,
@@ -25,6 +24,7 @@ from loki import (
     DeferredTypeSymbol, Module, HAVE_FP, FindExpressions, LiteralList, FindInlineCalls,
     AttachScopesMapper, FindTypedSymbols, Reference, Dereference
 )
+from loki.build import jit_compile, clean_test
 from loki.expression import symbols
 from loki.frontend import available_frontends, OFP, OMNI, FP
 from loki.tools import gettempdir, filehash
